db/requirements.txt
os
cv2
onnxruntime 
insightface 
faiss-cpu 
numpy 
opencv-python
pillow ### Changed from PIL to pillow
fastapi
uvicorn
